move functions to their own module
